mandatory part

+++ ft_dec_to_hex.c
@@ -6,20 +6,20 @@
 /*   By: sgoncalv <sgoncalv@student.42lisboa.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/23 10:42:53 by sgoncalv          #+#    #+#             */
-/*   Updated: 2023/02/02 08:32:14 by sgoncalv         ###   ########.fr       */
+/*   Updated: 2023/02/03 18:05:47 by sgoncalv         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-void ft_dec_to_hex(size_t dec)
+void	ft_dec_to_hex(unsigned int dec)
 {
-    char *base;
-    
-    base = "0123456789abcdef";
-    if ((dec / 16) > 0)
-    {
-        ft_dec_to_hex((dec / 16));
-    }
-    write(1, &base[dec % 16], 1);
+	char	*base;
+
+	base = "0123456789abcdef";
+	if ((dec / 16) > 0)
+	{
+		ft_dec_to_hex((dec / 16));
+	}
+	write(1, &base[dec % 16], 1);
 }
